Propagate all the needed interfaces and update data types

+++ c/src/server/server_replica.rs
@@ -50,8 +50,8 @@ pub extern "C" fn server_replica_drop(replica_info: *mut ServerReplica) {
 
 /// Returns the id of this replica.
 #[no_mangle]
-pub extern "C" fn server_replica_id(replica_info: *const ServerReplica) -> u64 {
-    borrow(replica_info).id()
+pub extern "C" fn server_replica_id(replica_info: *const ServerReplica) -> i64 {
+    borrow(replica_info).id() as i64
 }
 
 /// Returns the address this replica is hosted at.
@@ -75,5 +75,5 @@ pub extern "C" fn server_replica_is_primary(replica_info: *const ServerReplica) 
 /// Returns the raft protocol ‘term’ of this replica.
 #[no_mangle]
 pub extern "C" fn server_replica_term(replica_info: *const ServerReplica) -> i64 {
-    borrow(replica_info).term()
+    borrow(replica_info).term() as i64
 }
